backend/requirements.txt
# Django & MySQL
Django>=5.0,<6.0
mysqlclient>=2.2.0

# Django REST Framework + JWT Auth
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.1

# CORS & Env Config
django-cors-headers>=4.3.1
python-dotenv>=1.0.1
